refactor get_aws_env_vars
